requirements.txt
PySide6>=6.6
Pillow>=10.0
numpy>=2.0
rarfile>=4.0
PyMuPDF>=1.23
